add validation when try to save snapshot to dataverse without saving …

+++ seerDiscoverDataTableROMPCF/src/Utils/EstimateAveRate/custom.requirements.utils.ts
@@ -73,7 +73,7 @@ export const generateCustomRequirementMValue = async(inititlaData: any, conditio
     const {BaseData, resourceModelData, ModuleData, parameterModel, CustomisationModels, CustomRequirmentModel} = inititlaData
     let {hoursPerday} = parameterModel[0]
     if (hasParameters) {
-      hoursPerday = parseInt(settingParameters?.formattedData[
+      hoursPerday = parseFloat(settingParameters?.formattedData[
         parameterKeyIndex.hoursPerDay
       ]?.currentValue || '0');
     }
